Mark zeroconf.core as protected by renaming to zeroconf._core (#575)
