+++ zeroconf/__init__.py
@@ -23,7 +23,7 @@
 import sys
 
 from .cache import DNSCache  # noqa # import needed for backwards compat
-from .core import NotifyListener, Zeroconf  # noqa # import needed for backwards compat
+from ._core import NotifyListener, Zeroconf  # noqa # import needed for backwards compat
 from ._dns import (  # noqa # import needed for backwards compat
     DNSAddress,
     DNSEntry,
