+++ zeroconf/aio.py
@@ -26,9 +26,9 @@ import threading
 from types import TracebackType  # noqa # used in type hints
 from typing import Awaitable, Callable, Dict, List, Optional, Type, Union
 
+from ._core import NotifyListener, Zeroconf
 from ._dns import DNSOutgoing
 from .const import _BROWSER_TIME, _CHECK_TIME, _LISTENER_TIME, _MDNS_PORT, _REGISTER_TIME, _UNREGISTER_TIME
-from .core import NotifyListener, Zeroconf
 from .exceptions import NonUniqueNameException
 from .services import ServiceInfo, _ServiceBrowserBase, instance_name_from_service_info
 from .utils.aio import wait_condition_or_timeout
