+++ zeroconf/services/__init__.py
@@ -59,7 +59,7 @@ from ..utils.time import current_time_millis, millis_to_seconds
 
 if TYPE_CHECKING:
     # https://github.com/PyCQA/pylint/issues/3525
-    from ..core import Zeroconf  # pylint: disable=cyclic-import
+    from .._core import Zeroconf  # pylint: disable=cyclic-import
 
 
 @enum.unique
